+++ app/routes/($locale).products.$productHandle.tsx
@@ -1,9 +1,11 @@
+import type {
+  ShopifyAnalyticsProduct} from '@shopify/hydrogen';
 import {
   AnalyticsPageType,
-  getSelectedProductOptions,
-  ShopifyAnalyticsProduct,
+  getSelectedProductOptions
 } from '@shopify/hydrogen';
-import {defer, LoaderFunctionArgs} from '@shopify/remix-oxygen';
+import type { LoaderFunctionArgs} from '@shopify/remix-oxygen';
+import {defer} from '@shopify/remix-oxygen';
 import type {ProductRecommendationsQuery} from 'storefrontapi.generated';
 import invariant from 'tiny-invariant';
 import {routeHeaders} from '~/data/cache';
@@ -16,8 +18,9 @@ import {seoPayload} from '~/lib/seo.server';
 import type {Storefront} from '~/lib/type';
 import {WeaverseContent} from '~/weaverse';
 import {useLoaderData, useSearchParams} from '@remix-run/react';
-import {SelectedOptionInput} from '@shopify/hydrogen/storefront-api-types';
+import type {SelectedOptionInput} from '@shopify/hydrogen/storefront-api-types';
 import {useEffect} from 'react';
+import { getJudgemeReviews } from '~/lib/judgeme';
 
 export const headers = routeHeaders;
 
@@ -81,6 +84,13 @@ export async function loader({params, request, context}: LoaderFunctionArgs) {
     url: request.url,
   });
 
+  let judgeme_API_TOKEN = context.env.JUDGEME_PUBLIC_TOKEN;
+  let judgemeReviews = null;
+  if (judgeme_API_TOKEN) {
+    let shop_domain = context.env.PUBLIC_STORE_DOMAIN;
+    judgemeReviews = await getJudgemeReviews(judgeme_API_TOKEN, shop_domain, productHandle);
+  }
+
   return defer({
     variants,
     product,
@@ -95,6 +105,7 @@ export async function loader({params, request, context}: LoaderFunctionArgs) {
     },
     seo,
     weaverseData: await context.weaverse.loadPage(),
+    judgemeReviews
   });
 }
